refactor : SafeArea WebViewContainer로 이동
