+++ App.js
@@ -1,112 +1,15 @@
-import messaging from '@react-native-firebase/messaging';
-import React, {useEffect, useRef, useState} from 'react';
-import {
-  Alert,
-  AppRegistry,
-  BackHandler,
-  PermissionsAndroid,
-  Platform,
-  SafeAreaView,
-  StatusBar,
-} from 'react-native';
-import {PERMISSIONS, request} from 'react-native-permissions';
+import React from 'react';
+import {StatusBar, View} from 'react-native';
 
 import MainNavigator from './components/MainNavigator';
 
-// Firebase 알림
-async function requestUserPermission() {
-  const authStatus = await messaging().requestPermission();
-  const enabled =
-    authStatus === messaging.AuthorizationStatus.AUTHORIZED ||
-    authStatus === messaging.AuthorizationStatus.PROVISIONAL;
-
-  if (enabled) {
-    console.log('Authorization status:', authStatus);
-  }
-}
-
 export default function App() {
-  const webViewRef = useRef();
-
-  // 앱이 Foreground 인 상태에서 푸쉬알림 받는 코드 작성
-  useEffect(() => {
-    const unsubscribe = messaging().onMessage(async remoteMessage => {
-      Alert.alert('A new FCM message arrived!', JSON.stringify(remoteMessage));
-    });
-
-    return unsubscribe;
-  }, []);
-
-  // 앱이 Background 이거나 꺼진 상태에서 푸쉬알림 받는 코드 작성
-  messaging().setBackgroundMessageHandler(async remoteMessage => {
-    console.log('Message handled in the background!', remoteMessage);
-  });
-
-  AppRegistry.registerComponent('app', () => App);
-
-  // 뒤로 가기 control
-  const [exit, setexit] = useState(false);
-  const [swexit, setswexit] = useState(0);
-  const backAction = () => {
-    // 500(0.5초) 안에 back 버튼을 한번 더 클릭 할 경우 앱 종료
-    setswexit(swexit => swexit + 1);
-    return true;
-  };
-
-  useEffect(() => {
-    let timer;
-    if (exit === false) {
-      webViewRef?.current?.goBack();
-      setexit(true);
-      timer = setTimeout(function () {
-        setexit(false);
-      }, 500);
-    } else {
-      clearTimeout(timer);
-      BackHandler.exitApp();
-    }
-  }, [swexit]);
-
-  useEffect(() => {
-    BackHandler.addEventListener('hardwareBackPress', backAction);
-    return () =>
-      BackHandler.removeEventListener('hardwareBackPress', backAction);
-  }, []);
-
-  // 권한설정
-  useEffect(() => {
-    if (Platform.OS === 'android') {
-      const getper = async () => {
-        await PermissionsAndroid.request(
-          PermissionsAndroid.PERMISSIONS.ACCESS_MEDIA_LOCATION,
-        );
-        await PermissionsAndroid.request(PermissionsAndroid.PERMISSIONS.CAMERA);
-      };
-      getper();
-    }
-    if (Platform.OS === 'ios') {
-      const requestPhotoLibraryPermission = async () => {
-        await request(PERMISSIONS.IOS.CAMERA);
-        await request(PERMISSIONS.IOS.MICROPHONE);
-        await request(PERMISSIONS.IOS.PHOTO_LIBRARY);
-        await requestUserPermission();
-      };
-      requestPhotoLibraryPermission();
-    }
-  }, []);
-
   return (
-    <SafeAreaView
-      overScrollMode="never"
-      style={{flex: 1, backgroundColor: 'white'}}>
-      <StatusBar
-        animated={false}
-        backgroundColor="white"
-        translucent={false}
-        hidden={false}
-        barStyle="dark-content"
-      />
-      <MainNavigator />
-    </SafeAreaView>
+    <>
+      <StatusBar />
+      <View style={{flex: 1}}>
+        <MainNavigator />
+      </View>
+    </>
   );
 }
